Refactor: Reorganizing files related to the role and user
