+++ src/main/java/com/ecommerceproject/modules/user/entity/User.java
@@ -1,19 +1,15 @@
 package com.ecommerceproject.modules.user.entity;
 
-import com.ecommerceproject.domain.model.Role;
 import jakarta.persistence.*;
 import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
-import java.util.Optional;
 
 @Entity
 @Table(name = "TB_USER", schema = "EXCHANGE_APP")
@@ -75,10 +71,5 @@ public class User implements UserDetails {
     @Override
     public boolean isEnabled() {
         return true;
-    }
-
-    @Repository
-    public static interface RoleRepository extends CrudRepository<Role, Integer> {
-        Optional<Role> findByName(String name);
     }
 }
