+++ src/main/java/com/ecommerceproject/modules/user/repositories/RoleRepository.java
@@ -1,12 +1,8 @@
 package com.ecommerceproject.modules.user.repositories;
 
-import com.ecommerceproject.domain.model.Role;
-import org.springframework.data.repository.CrudRepository;
+import com.ecommerceproject.modules.user.entity.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Integer> {
-    Optional<Role> findByName(String name);
-}
+public interface RoleRepository extends JpaRepository<Role, Integer> {}
