+++ src/main/java/com/ecommerceproject/modules/user/repositories/UserRepository.java
@@ -1,12 +1,12 @@
 package com.ecommerceproject.modules.user.repositories;
 
 import com.ecommerceproject.modules.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository {
-    List<User> findAll();
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-    User save(User user);
 }
